api/requirements.txt
fastapi==0.136.1
uvicorn==0.46.0
langchain==1.2.15
langchain-groq==1.1.2
langgraph==1.1.9
python-dotenv==1.2.2
pandas==3.0.2
yfinance==1.3.0
aiosqlite==0.21.0
